Refactor: moved 'auth' folder from cmd to pkg
