+++ cmd/uid_provider/client/interface.go
@@ -3,7 +3,7 @@ package client
 import (
 	"context"
 
-	"gitlab.inspr.dev/inspr/core/cmd/insprd/auth"
+	"gitlab.inspr.dev/inspr/core/pkg/api/auth"
 )
 
 // User defines the information a user contains
