+++ pkg/controller/mocks/auth.go
@@ -3,7 +3,7 @@ package mocks
 import (
 	"context"
 
-	"gitlab.inspr.dev/inspr/core/cmd/insprd/auth"
+	"gitlab.inspr.dev/inspr/core/pkg/api/auth"
 	"gitlab.inspr.dev/inspr/core/pkg/controller"
 )
